+++ example/lib/presentation/src/widgets/src/calendar/calendar.dart
@@ -1,2 +1,4 @@
+export 'src/ac_calendar_child_delegate.dart';
 export 'src/ac_calendar_layout.dart';
+export 'src/ac_calendar_month_cache.dart';
 export 'src/ac_calendar_widget.dart';