+++ example/lib/presentation/src/widgets/src/calendar/src/ac_calendar_month_cache.dart
@@ -1,0 +1,12 @@
+import '../../../../../presentation.dart';
+
+/// Кэшированные данные месяца
+class ACCalendarMonthCache {
+  const ACCalendarMonthCache({
+    required this.days,
+    required this.layout,
+  });
+
+  final List<DateTime> days;
+  final DefaultMonthLayout layout;
+}